test with double type

+++ testQueue.c
@@ -2,45 +2,45 @@
 #include <stdio.h>
 
 int main(void) {
-    queue myqueue = newQueue(int);
-    int input = 0, out = 0;
+    queue myqueue = newQueue(double);
+    double input = 0, out = 0;
     printf("Push 5> ");
     for (int i = 0; i < 5; ++i) {
-        scanf("%d", &input);
+        scanf("%lf", &input);
         myqueue.push(&myqueue, &input);
     }
     printf("Pop 4> ");
     while (myqueue.size - 1) {
         myqueue.pop(&myqueue, &out);
-        printf("%d ", out);
+        printf("%g ", out);
     }
     putchar('\n');
     printf("Push 3> ");
     for (int i = 0; i < 3; ++i) {
-        scanf("%d", &input);
+        scanf("%lf", &input);
         myqueue.push(&myqueue, &input);
     }
     printf("Front> ");
-    printf("%d\n", *(int *) myqueue.front(&myqueue));
+    printf("%g\n", *(double *) myqueue.front(&myqueue));
     printf("Back> ");
-    printf("%d\n", *(int *) myqueue.back(&myqueue));
+    printf("%g\n", *(double *) myqueue.back(&myqueue));
     myqueue.clear(&myqueue);
     if (myqueue.empty(&myqueue)) {
         puts("Now the stack is empty");
     }
     printf("Push 3> ");
     for (int i = 0; i < 3; ++i) {
-        scanf("%d", &input);
+        scanf("%lf", &input);
         myqueue.push(&myqueue, &input);
     }
     printf("Front> ");
-    printf("%d\n", *(int *) myqueue.front(&myqueue));
+    printf("%g\n", *(double *) myqueue.front(&myqueue));
     printf("Back> ");
-    printf("%d\n", *(int *) myqueue.back(&myqueue));
+    printf("%g\n", *(double *) myqueue.back(&myqueue));
     printf("Pop 3> ");
     while (myqueue.size) {
         myqueue.pop(&myqueue, &out);
-        printf("%d ", out);
+        printf("%g ", out);
     }
     myqueue.destory(&myqueue);
     return 0;
